app/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
itsdangerous==2.2.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.141.0
pytz==2024.1
requests==2.32.3
Flask-Talisman==1.1.0
gunicorn==22.0.0
azure-cosmos==4.7.0
sendgrid==6.11.0
